==> jupyter_notebooks/1a_R_script_transcriptomic_to_genomic_coordinates.ipynb <==
# Theoretical NTE

library(dplyr)
library(GenomicFeatures)

# this scripts transform local (transcriptomic) coordinates of extensions into global (genomic) coordinates using .gtf file. 
txdb <- makeTxDbFromGFF('data/gencode.v25.annotation.gtf')
cr <- read.csv('tmp_res/local_coo_primary_NTE.txt',  sep='\t', stringsAsFactors = FALSE)

tr_names <- cr$tr_id
start <- cr$N_term_start1 + 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$N_term_end1

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/primary_NTE_global_coo.txt', sep='\t')

# 50-codons of NTE 

cr <- read.csv('tmp_res/local_coo_50codons_and_less.txt',  sep='\t', stringsAsFactors = FALSE)

tr_names <- cr$tr_id
start <- cr$X50len_flag + 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$N_term_end1 

print (length(tr_names))
print (length(start))
print (length(end))

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/global_coo_50codons_and_less.txt', sep='\t')

# Gencode25: coding exons 

cr <- read.csv('tmp_res/metadata_pc_g25.txt',  sep='\t', stringsAsFactors = FALSE)

tr_names <- cr$tr_id
start <- cr$cds_start_pos + 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$cds_stop_pos 

print (length(tr_names))
print (length(start))
print (length(end))

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/coding_exons_g25_global.txt', sep='\t')

# Gencode35: coding exons 

txdb_g35 <- makeTxDbFromGFF('data/gencode.v35.annotation.gtf')

cr <- read.csv('tmp_res/metadata_pc_g35.txt',  sep='\t', stringsAsFactors = FALSE)

tr_names <- cr$tr_id
start <- cr$cds_start_pos + 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$cds_stop_pos 

print (length(tr_names))
print (length(start))
print (length(end))

exons <- exonsBy(txdb_g35, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/coding_exons_g35_global.txt', sep='\t')

# PCSF per codon: global coordinates of codons in NTE+50 codons of CDS

### PhyloSET

txdb <- makeTxDbFromGFF('data/gencode.v25.annotation.gtf')
cr <- read.csv('tmp_res/local_coo_of_codons_for_PCSF_plot_PhyloSET_plus50codonsCDS.txt',  sep='\t', stringsAsFactors = FALSE)

tr_names <- cr$tr_id
start <- cr$codon_start+ 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$codon_end

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/global_coo_of_codons_for_PCSF_plot_PhyloSET_plus50codonsCDS.txt', sep='\t')

### RiboSET

cr <- read.csv('tmp_res/local_coo_of_codons_for_PCSF_plot_RiboSET_plus50codonsCDS.txt',  sep='\t', stringsAsFactors = FALSE)


tr_names <- cr$tr_id
start <- cr$codon_start+ 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$codon_end

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/global_coo_of_codons_for_PCSF_plot_RiboSET_plus50codonsCDS.txt', sep='\t')

# CYTH2: NTE+50codons of CDS 

tr_names <- c('ENST00000452733.6')
start <- c(302+1)
end <- c(626)

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))

out %>% filter(hit == TRUE)

chr19:48469334-48469526+chr19:48470353-48470483

# MARVELD1: NTE+50codons of CDS 

txdb <- makeTxDbFromGFF('data/gencode.v25.annotation.gtf')

tr_names <- c('ENST00000285605.7')
start <- c(27+1)
end <- c(303)

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))

out %>% filter(hit == TRUE)

chr10:97713751-97714026, +

# ARPC1A

tr_names <- c('ENST00000262942.9')
start <- c(1+1)
end <- c(274)

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))

out %>% filter(hit == TRUE)

chr7:99325911-99326004+chr7:99333325-99333417+chr7:99338181-99338266

# SPFQ: NTE+50codons of CDS 

tr_names <- c('ENST00000357214.5')
start <- c(0+1)
end <- c(99+50*3)

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))

out %>% filter(hit == TRUE)

chr1:35192900-35193148, -

# Reproducibility of PhyloCSF-based approach on 24/59 genes from Ivanov et al 2011 paper 

txdb_g35 <- makeTxDbFromGFF('data/gencode.v35.annotation.gtf')


cr <- read.csv('tmp_res/reprod_24genes_2011_local_coo_df.txt',  sep='\t', stringsAsFactors = FALSE)

tr_names <- cr$tr_id
start <- cr$X50len_flag + 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$new_N_term_end 

print (length(tr_names))
print (length(start))
print (length(end))

exons <- exonsBy(txdb_g35, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/reprod_24genes_2011_global_coo_df.txt', sep='\t')

# Predicted starts&ext in RiboSET

txdb_g25 <- makeTxDbFromGFF('data/gencode.v25.annotation.gtf')

cr <- read.csv('tmp_res/RiboSET_local_coo_pred_start.txt',  sep='\t', stringsAsFactors = FALSE)

tr_names <- cr$tr_id
start <- cr$start_codon_local_coo_start #it was zero  # cds_start_pos	utr5_start
end <- cr$ start_codon_local_coo_stop-1

print (length(tr_names))
print (length(start))
print (length(end))

exons <- exonsBy(txdb_g25, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/RiboSET_global_coo_pred_start.txt', sep='\t')

tr_names <- cr$tr_id
start <- cr$start_codon_local_coo_start #it was zero  # cds_start_pos	utr5_start
end <- cr$N_term_end1

print (length(tr_names))
print (length(start))
print (length(end))

exons <- exonsBy(txdb_g25, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/RiboSET_global_coo_pred_start_ext.txt', sep='\t')

# Codons coo (riboset) for COSMIC variants 

# this scripts transform local (transcriptomic) coordinates of extensions into global (genomic) coordinates using .gtf file. 
txdb <- makeTxDbFromGFF('data/gencode.v25.annotation.gtf')
cr <- read.csv('tmp_res/ext_codons_local_coo_cosmic.txt',  sep='\t', stringsAsFactors = FALSE)

tr_names <- cr$tr_id
start <- cr$codon_start + 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$codon_stop

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]

x <- GRanges(tr_names, IRanges(start=start, end=end)) #, 

out <- data.frame(pmapFromTranscripts(x, exons))

write.table(out, 'tmp_res/ext_codons_global_coo_cosmic.txt', sep='\t')

out

# mice CDS v14 

# this scripts transform local (transcriptomic) coordinates of extensions into global (genomic) coordinates using .gtf file. 
txdb <- makeTxDbFromGFF('data/gencode.vM14.annotation.gtf')


cr <- read.csv('tmp_res/mice14_nte_riboset.txt',  sep='\t', stringsAsFactors = FALSE)

tr_names <- cr$tr_id
start <- cr$start + 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$stop

length(tr_names)
length(start)
length(end)

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]


x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/mice14_nte_riboset_global_coo.txt', sep='\t')

# mouse CDS

cr <- read.csv('tmp_res/mice14_CDS_local_coo.txt',  sep='\t', stringsAsFactors = FALSE)
head(cr)

tr_names <- cr$tr_id
start <- cr$cds_start_pos + 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$cds_stop_pos

length(tr_names)
length(start)
length(end)

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/mice14_CDS_global_coo.txt', sep='\t')

# Gencode annotation genes 

# this scripts transform local (transcriptomic) coordinates of extensions into global (genomic) coordinates using .gtf file. 
txdb <- makeTxDbFromGFF('data/gencode.v25.annotation.gtf')

cr <- read.csv('tmp_res/human_genes_for_annotation_in_gencode_local_coo.txt',  
               sep='\t', stringsAsFactors = FALSE)

head(cr)

tr_names <- cr$tr_id
start <- cr$new_cds_start + 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$cds_stop

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/human_genes_for_annotation_in_gencode_global_coo.txt', sep='\t')

# Genes in mice 

# 

# this scripts transform local (transcriptomic) coordinates of extensions into global (genomic) coordinates using .gtf file. 
txdb <- makeTxDbFromGFF('data/gencode.vM14.annotation.gtf')

cr <- read.csv('tmp_res/mouse_genes_for_annotation_local_coo.txt',  
               sep='\t', stringsAsFactors = FALSE)

head(cr)

tr_names <- cr$mouse.tr_id
start <- cr$mouse_new_cds_start + 1 #it was zero  # cds_start_pos	utr5_start
end <- cr$mouse_cds_stop

length(tr_names)
length(start)
length(end)

exons <- exonsBy(txdb, by="tx", use.names=TRUE)[tr_names]
x <- GRanges(tr_names, IRanges(start=start, end=end)) #,  
out <- data.frame(pmapFromTranscripts(x, exons))
write.table(out, 'tmp_res/mouse_genes_for_annotation_global_coo.txt', sep='\t')
