==> jupyter_notebooks/2_v2_Adding_Riboseq_Proteomics_number_of_records_PhyloSET_and_RiboSET.ipynb <==
%matplotlib inline

import numpy as np
import collections
from collections import OrderedDict, Counter, defaultdict
import pandas as pd

import Bio
from Bio import SeqIO
from Bio.Seq import Seq
from Bio.SeqUtils import molecular_weight
from Bio.Alphabet import generic_dna, generic_rna, generic_protein

import seaborn as sns
import matplotlib.pyplot as plt

import glob

import subprocess
from subprocess import call

import re

import pickle

import sys                                                                         
from gffutils.iterators import DataIterator  
from gtfparse import read_gtf

# Open metadata files 

# gencode 25 metadata 
metadata_pc_g25 = pd.read_csv('tmp_res/metadata_pc_g25.txt', sep='\t')

# gencode 35 metadata 
metadata_pc_g35 = pd.read_csv('tmp_res/metadata_pc_g35.txt', sep='\t')

# open metadata with scores and metrics
meta = pd.read_csv('tmp_res/local_and_global_df_g25ovlp_g35ovlp_g38ovlp_refseqovlp_strand_FRAMES_PhyloCSF.txt', sep='\t')

# genes from 2011 paper
genes_42 = 'RASL10B,FGFR1,YPEL2,ENOX2,UFSP1,WDR26,TIAL1,YPEL4,NGF,TRPV6,KCTD11,STARD10,YPEL1,R3HCC1,ZFP62,NHLRC4,EPHX3,C1QL4,TLE3,HDGF,ANKRD42,FAM217B,NFKBID,PTEN,RNF187,METTL23,TMEM8B,VANGL2,FNDC5,C1QL1,CITED1,EIF4G3,HELZ2,CYTH2,NTF3,MFSD4B,C1QL3,KCNN4,IFT46,RASD2,CITED2,C1QL2'.split(',')
genes_17 = 'GTF3A, EIF4G2, SP3, PRPS1L1, TEAD4, TEAD3, CACNG8, OAZ3, TEAD1, DDX17, VEGFA, NR1I2, HCK, WT1, BAG1, NPW, MYC'.split(', ')

# Add CDS completness and biotype='protein_coding' before selection SETs (gencode25)

# returns GTF with essential columns such as "feature", "seqname", "start", "end"
# alongside the names of any optional keys which appeared in the attribute column
df = read_gtf("data/gencode.v25.annotation.gtf")

# filter DataFrame to gene entries on chrY
df_genes = df[df["feature"] == "gene"]


df_CDS = df[df["feature"] == "CDS"]

annotation_genes = df_genes[['transcript_id', 'transcript_type', 'gene_name', 'tag']]
annotation_CDS = df_CDS[['transcript_id', 'transcript_type', 'gene_name', 'tag']]

annotation_CDS.columns = ['tr_id', 'transcript_type', 'gene', 'tag']

print (annotation_CDS.shape, annotation_genes.shape)

meta1 = meta.merge(annotation_CDS, on=['tr_id', 'gene'], how='inner').drop_duplicates()

meta1.shape, meta.shape

meta1[0:2]

# Add predictions from Trips-Viz (ranks, proteomics support)

transl_with_all_codons = pd.read_csv('data/extensions_TripsViz_1perGene_AUG_NearCognates.txt', sep=',')

transl_with_all_codons.columns = ['gene_ext', 'tr_id1', 'Start_ext', 'Stop_ext', 'len_ext', 
                                 'global_rank_ext', 'type_all_codons', 'trips_viz_link', 'Start_codon_ext', 
                                 'Highframe rank', 'Highframe value',
       'Lowframe rank', 'Lowframe value', 'Stop rank', 'Stop value',
       'Start rank', 'Start value', 'Coverage rank', 'Coverage_value_ext',
       'Amino acid sequence', 'Proteomics_count_ext', 'CDS_ratio']

transl_with_all_codons2 = transl_with_all_codons[['gene_ext', 'tr_id1', 'Start_ext', 'Stop_ext', 
                                                 'global_rank_ext', 'Start_codon_ext', 
                                                 'Proteomics_count_ext', 'Coverage_value_ext', 'CDS_ratio']]

transl_with_all_codons2['Riboseq_Summary'] = transl_with_all_codons2['Start_codon_ext']+'; '+transl_with_all_codons2['Start_ext'].astype(str)+'-'+transl_with_all_codons2['Stop_ext'].astype(str)+'; Rank: '+transl_with_all_codons2['global_rank_ext'].astype(str)+'; cov: '+np.round(transl_with_all_codons2['Coverage_value_ext'],2).astype(str)


transl_with_all_codons2[0:5]

meta_cov_riboseq_pred_proteomics_pred = meta1.merge(transl_with_all_codons2, on='tr_id1', how='left')

meta_cov_riboseq_pred_proteomics_pred['Nstart_Nend'] = meta_cov_riboseq_pred_proteomics_pred['N_term_start1'].astype(str)+'-'+meta_cov_riboseq_pred_proteomics_pred['N_term_end1'].astype(str)

DF = meta_cov_riboseq_pred_proteomics_pred.copy()

print (meta_cov_riboseq_pred_proteomics_pred.shape, meta.shape)

DF.to_csv('tmp_res/METADATA_CURRENT.txt', sep='\t', index=False)

DF[DF['gene'] == 'DHFR'][['tr_id', 'ovlp', 'ovlp_strandless']]

DF[DF['gene'] == 'DYRK1B'][['PhyloCSF120score', 'Riboseq_Summary']]

transl_with_all_codons[transl_with_all_codons['gene_ext'] == 'DYRK1B']

# PhyloSET 

* protein-coding biotype
* tags: complete CDS 
* no overlaps with coding exons 
* primary extension has at least 20 codons
* number of records >= 3
* PhyloCSF score > 0

DF = DF[DF['tag'].notna()]

* old version of SET1: we don't care about overlaps with coding exons on other strands = 60 genes vs 55 genes when we care

SET1_ = DF[(DF['PhyloCSF120score'] > 0) & (DF['ovlp'] == '0;0;0') & (DF['len_codons'] >= 20) &
   (DF['transcript_type'] == 'protein_coding') & (~DF['tag'].str.contains('cds_start_NF')) & 
           (~DF['tag'].str.contains('cds_end_NF')) &
    (DF['number_of_records'] >= 3) &  (DF['gene'] != 'SKI')]

SET1 = DF[(DF['PhyloCSF120score'] > 0) & 
   (DF['ovlp_strandless'] == '0;0;0') &
   (DF['len_codons'] >= 20) &
   (DF['transcript_type'] == 'protein_coding') &
                     (~DF['tag'].str.contains('cds_start_NF')) &
                     (~DF['tag'].str.contains('cds_end_NF')) &
    (DF['number_of_records'] >= 3) & 
    (DF['gene'] != 'SKI')]


SET1.gene.nunique()

DF[DF['tr_id'].isin(set(SET1_.tr_id.tolist()) - set(SET1.tr_id.tolist()))][['tr_id', 'gene', 'ovlp_strandless', 'global_coo_primary', 'strand', 'Riboseq_Summary', 'PhyloCSF120score']]

OAZ3 (ovlp=71nt, PCSF=93), AC007906.1 (ovlp=89nt, Rank=5040;cov=0, PCSF=70), GLOD4 (ovlp=61nt, PCSF=261),GDF5OS (ovlp=77nt, PCSF=1079, Rank=5739;cov=0), RANBP1 (ovlp=83nt, PCSF=310)

SET1_[['tr_id', 'gene', 'N_term_start1','N_term_end1', 'len_codons', 'global_coo_50_and_less', 'strand', 'global_coo_primary', 'PhyloCSF120score', 'number_of_records', 'ovlp', 'Riboseq_Summary', 
                    'Coverage_value_ext', 'Proteomics_count_ext', 'tag', 'transcript_type']].to_csv('tmp_res/SET1.txt', sep='\t', index=False)



# RiboSET

* protein-coding biotype
* tags: complete CDS 
* no overlaps with coding exons 
* primary extension has at least 20 codons
* rank is less than 500 

SET2_ = DF[(DF['transcript_type'] == 'protein_coding') &
                     (~DF['tag'].str.contains('cds_start_NF')) &
                     (~DF['tag'].str.contains('cds_end_NF')) & (DF['ovlp'] == '0;0;0') & 
                    (DF['global_rank_ext'] <= 500) & 
                    (DF['len_codons'] >= 20)]

print (SET2_.gene.nunique(), SET2_.tr_id.nunique())

SET2 = DF[(DF['transcript_type'] == 'protein_coding') &
                     (~DF['tag'].str.contains('cds_start_NF')) &
                     (~DF['tag'].str.contains('cds_end_NF')) & (DF['ovlp_strandless'] == '0;0;0') & 
                    (DF['global_rank_ext'] <= 500) & 
                    (DF['len_codons'] >= 20)]

print (SET2.gene.nunique(), SET2.tr_id.nunique())

* сheck if there is an AUG-extension 

SET2_[~SET2_['Riboseq_Summary'].str.contains('ATG')][['gene', 'tr_id', 'Riboseq_Summary', 'ovlp', 'ovlp_strandless']].gene.nunique()

# genes with 2 transcripts: onc chrX and chrY chromosomes 
SET2[SET2['gene'].isin(['SLC25A6', 'CD99'])][['tr_id', 'gene', 'global_coo_50_and_less']]

DF[DF['tr_id'].isin(set(SET2_.tr_id.tolist()) - set(SET2.tr_id.tolist()))][['tr_id', 'gene', 'ovlp_strandless', 'global_coo_primary', 'strand', 'Riboseq_Summary', 'PhyloCSF120score']]

TRIP13 (ovlp=12nt, Rank=107, score:-1324);
DHFR (ovlp=237nt, Rank=291, score=-2881)
SLC25A32 (ovlp=109nt except RefSeq=0, Rank=64, score=-985);
NIP7 (ovlp=121nt/202nt, Rank=382, score=-2914)

# save it 
SET2_[['tr_id', 'gene', 'N_term_start1','N_term_end1', 'len_codons', 'global_coo_50_and_less', 'strand', 'global_coo_primary', 'PhyloCSF120score', 'number_of_records', 'ovlp', 'Riboseq_Summary', 
                    'Coverage_value_ext', 'Proteomics_count_ext', 'tag', 'transcript_type', 'CDS_ratio']].to_csv('tmp_res/SET2.txt', sep='\t', index=False)


DF[DF['gene'] == 'DHFR'][['tr_id', 'gene', 'Riboseq_Summary', 'ovlp', 'ovlp_strandless']]

SET2_[SET2_['gene'] == 'DHFR']

SET2[SET2['gene'] == 'DHFR']

# Genes that are supported by proteomics 

SET1_[SET1_['Proteomics_count_ext'] > 0].gene.nunique()

SET2_[SET2_['Proteomics_count_ext'] > 0].gene.nunique()

18/PhyloSET, 91/RiboSET

', '.join(list(set(SET1_.gene.tolist()).intersection(SET2_.gene.tolist())))

SET1_.gene.nunique(), SET2_.gene.nunique()

# ATG-extensions in RiboSET

SET2_[SET2_['Start_codon_ext'] == 'ATG']

SET2_[SET2_['Start_codon_ext'] != 'ATG'].tr_id.nunique()

SET1_.tr_id.nunique(), SET1_.gene.nunique()

# Add info about in-frame ATG in theoretical NTE 

for tr_id in list(set(SET1_.tr_id.tolist()+SET2_.tr_id.tolist())):
    #print (tr_id)
    # extract primary NTE sequence
    N_term_start = meta[meta['tr_id'] == tr_id].iloc[0].N_term_start1 
    N_term_end = meta[meta['tr_id'] == tr_id].iloc[0].N_term_end1
    NTE_seq = metadata_pc_g25[metadata_pc_g25['tr_id'] == tr_id].iloc[0].transcript_seq[N_term_start:N_term_end]
    
    codons = [NTE_seq[i:i+3] for i in range(0, len(NTE_seq), 3)]
    atgs = []
    for i, codon in enumerate(codons):
        if codon == 'ATG':
            atgs.append('ATG_'+str(i*3))
    
    if len(atgs) != 0:
        print ( metadata_pc_g25[metadata_pc_g25['tr_id'] == tr_id].iloc[0].gene, tr_id, '; '.join(atgs))
