==> jupyter_notebooks/3b_v2_Comparison_with_other_studies.ipynb <==
%matplotlib inline

import numpy as np
import collections
from collections import OrderedDict, Counter, defaultdict
import pandas as pd

import Bio
from Bio import SeqIO

import seaborn as sns
import matplotlib.pyplot as plt

import glob

import subprocess
from subprocess import call

import re

import pickle

from gtfparse import read_gtf

# PMID: 24623590

ext_df = pd.read_csv('data/Lee_PMC4014282_extensions.txt', sep='\t', header=None)

ext_df = ext_df[[12, 13, 14, 8, 9, 16]]

ext_df.columns = ['tr_id', 'gene', 'start_codon', 'peptide', 'pos_relative_to_TIS', 'type']

ext_df[ext_df['type'] == 'n-term-ext'].to_csv('plots/final_S/SupY_PMC4014282_17extensions_Nterminomics_Riboseq.txt', sep='\t', index=False)

ext_df.tr_id.tolist()

# Compare with PhyloSET and RiboSET

# open SET1 (PhyloSET) and SET2 (RiboSET)
SET1 = pd.read_csv('tmp_res/SET1.txt', sep='\t')
SET2 = pd.read_csv('tmp_res/SET2.txt', sep='\t')

', '.join(list(set(SET1.gene.tolist()).intersection(ext_df.gene.tolist())))

', '.join(list(set(SET2.gene.tolist()).intersection(ext_df.gene.tolist())))

Two genes from PhyloSET (FXR2, HNRNPA0) and seven genes from RiboSET (NARS, HDGF, HNRNPA0, FXR2, SYAP1, KAT7, BAG6) were found in 

ext_df

# Comprehensive analysis of human protein N-termini enables assessment of various protein forms


# open metadata with scores and metrics
meta = pd.read_csv('tmp_res/METADATA_CURRENT.txt', sep='\t')

# gencode 25
metadata_pc_g25 = pd.read_csv('tmp_res/metadata_pc_g25.txt', sep='\t')


nterm_df = pd.read_csv('data/PMID28747677_nterm.txt', sep='\t', header=None)

nterm_df.columns = ['chr', 'start', 'strand', 'tr_id1', 
                   'uniprot_id', 'pos_to_CDS', 'aa_seq', 'nt_seq']

print (nterm_df.shape[0], nterm_df.tr_id1.nunique())

nterm_df[0:4]

metadata_pc_g25['tr_id1'] = [x.split('.')[0] for x in metadata_pc_g25['tr_id'].tolist()]

nterm_df2 = nterm_df.merge(metadata_pc_g25[['tr_id1', 'gene']], on='tr_id1', how='left')

nterm_df2.shape[0], nterm_df2.drop_duplicates().shape[0]

nterm_genes = list(nterm_df2[nterm_df2['pos_to_CDS'] <= -3].gene.unique())

print (len(nterm_genes))

# phyloset 

print (len(set(SET1.gene.tolist()).intersection(nterm_genes))), set(SET1.gene.tolist()).intersection(nterm_genes)

# riboset

print (len(set(SET2.gene.tolist()).intersection(nterm_genes)))

We also compared 

PMID: 28747677 

Only 1 gene from PhyloSET and 17 genes from RiboSET were present among 171 genes with N-terminal extensions detected by N-terminomics. 

nterm_df2

nterm_df2[nterm_df2['gene'].isin(SET1.gene.tolist()+SET2.gene.tolist())][['tr_id1', 'pos_to_CDS', 'gene', 'nt_seq']].to_csv('tmp_res/comparison_with_PMID28747677.txt', sep='\t', index=False)
