==> jupyter_notebooks/4_v2_GenCode_and_RefSeq_db_screening.ipynb <==
%matplotlib inline

import numpy as np
import collections
from collections import OrderedDict, Counter, defaultdict
import pandas as pd

import Bio
from Bio import SeqIO

import seaborn as sns
import matplotlib.pyplot as plt

import glob

import subprocess
from subprocess import call

import re

import pickle

import sys                                                                         
from gffutils.iterators import DataIterator  
from gtfparse import read_gtf

# Refseq v109: find all non-AUG proteoforms, with sole start and additional forms 

* using web-interface, extract genbank file with genes containing non-AUG in their description 
* parse it, get a gene list 

# new_refseq_nonAUG.gb

gene_names_with_non_AUG_starts = []

gb_file = "data/RefSeq_nonAUG_genes_May2021.gb"
for i, gb_record in enumerate(SeqIO.parse(open(gb_file,"r"), "genbank")) :
    # now do something with the record
    
    print ('record %s / 146' % i)
    # print ("Name %s, %i features" % (gb_record.name, len(gb_record.features)))
    for feat in gb_record.features:
        try:
            if feat.type == 'CDS':
                non_AUG = None
                cds_seq = str(feat.extract(gb_record.seq))
                non_AUG_from_cds_seq = cds_seq[0:3] 
                stop_codon = cds_seq[-3:]
                gene_name = feat.qualifiers['gene'][0]
                for el in feat.qualifiers['note']:
                    if el.find('non-AUG') != -1:
                        non_AUG = el.split('non-AUG')[1].split(' ')[1].replace('(', '').replace(')', '')
                #li.append([gb_record.name, gene_name, str(gb_record.seq), non_AUG, cds_seq, non_AUG_from_cds_seq, stop_codon])
                        gene_names_with_non_AUG_starts.append(gene_name)
                
                
        except Exception as e:
            print (gb_record.name, e)


print ('%s genes with non-AUG proteoforms annotated in refseq' % len(set(gene_names_with_non_AUG_starts)))
print (' | '.join(list(set(gene_names_with_non_AUG_starts))))

* download ALL genbank files for 43 selected genes (including all proteoforms: AUG and nonAUG)

* we can use following query:

print ('('+'[All Fields] OR '.join(list(set(gene_names_with_non_AUG_starts)))+')')

((MORN2[All Fields] OR SWI5[All Fields] OR TEAD1[All Fields] OR MRTFA[All Fields] OR KCTD11[All Fields] OR STIM2[All Fields] OR WDR26[All Fields] OR C16orf82[All Fields] OR PIGX[All Fields] OR TEAD3[All Fields] OR EIF4G2[All Fields] OR RNF187[All Fields] OR TEAD4[All Fields] OR PEG10[All Fields] OR FNDC5[All Fields] OR TRPV6[All Fields] OR FGF2[All Fields] OR LUZP6[All Fields] OR DDX17[All Fields] OR BAG1[All Fields] OR HCK[All Fields] OR SP3[All Fields] OR HMHB1[All Fields] OR PTEN[All Fields] OR CACNG8[All Fields] OR MDFIC[All Fields] OR NPW[All Fields] OR IRAG1[All Fields] OR VEGFA[All Fields] OR GTF3A[All Fields] OR RARB[All Fields] OR NR1I2[All Fields] OR CEBPA[All Fields] OR PRPS1L1[All Fields] OR WT1[All Fields] OR PIM1[All Fields] OR OAZ3[All Fields] OR PANK2[All Fields] OR TXNRD3[All Fields] OR POU5F1[All Fields] OR MYC[All Fields] OR R3HCC1[All Fields] OR NPTXR[All Fields] OR NELFB)) AND "Homo sapiens"[porgn:__txid9606] 

li3 = []

gb_file = "data/RefSeq_genes_with_AUG_and_nonAUG_May2021.gb"
for gb_record in SeqIO.parse(open(gb_file,"r"), "genbank") :
    # now do something with the record
    # print ("Name %s, %i features" % (gb_record.name, len(gb_record.features)))
    for feat in gb_record.features:
        try:
            if (feat.type == 'CDS') & (gb_record.name.find('NM_') != -1):
                non_AUG = None
                cds_seq = str(feat.extract(gb_record.seq))
                non_AUG_from_cds_seq = cds_seq[0:3] 
                stop_codon = cds_seq[-3:]
                gene_name = feat.qualifiers['gene'][0]
                for el in feat.qualifiers['note']:
                    if el.find('non-AUG') != -1:
                        non_AUG = el.split('non-AUG')[1].split(' ')[1].replace('(', '').replace(')', '')
                li3.append([gb_record.name, gene_name, str(gb_record.seq), non_AUG, cds_seq, non_AUG_from_cds_seq, stop_codon])
                
        except Exception as e:
            print (gb_record.name, e)

Check which ones have only non-AUG proteoforms and which ones have both

refseq_df = pd.DataFrame(li3)
refseq_df.columns = ['tr_id', 'gene', 'transcript_seq', 'non_AUG_start', 'cds_seq', 'non_AUG_from_cds_seq', 'stop_codon']

d2 = refseq_df[['tr_id', 'gene', 'non_AUG_from_cds_seq']].groupby('gene')['non_AUG_from_cds_seq'].apply(set).apply(list).apply(', '.join).to_dict()
df = pd.DataFrame(zip(d2.keys(), d2.values()), columns = ['gene', 'codons'])
df = df[df['codons'] != 'ATG']


print ('%s genes have only non-AUG proteoroms' % df[(df['codons'].str.len() == 3) & (df['codons'] != 'ATG')].gene.nunique())
print (', '.join(df[(df['codons'].str.len() == 3) & (df['codons'] != 'ATG')].gene.tolist()))
print ()

print ('%s genes have both AUG and non-AUG proteoroms' % df[df['codons'].str.len() >= 6].gene.nunique())
print (', '.join(df[df['codons'].str.len() >= 6].gene.tolist()))
print ()

refseq_df['non_AUG_start'] = refseq_df['non_AUG_start'].str.replace('U', 'T')
refseq_only_non_AUG  = refseq_df[refseq_df['gene'].isin(df[(df['codons'].str.len() == 3) & (df['codons'] != 'ATG')].gene.unique())][['tr_id', 'gene', 'non_AUG_start']]
refseq_only_non_AUG.to_csv('tmp_res/refseq109_genes_with_sole_non_AUG.txt', sep='\t', index=False)

refseq_both = refseq_df[refseq_df['gene'].isin(df[df['codons'].str.len() >= 6].gene.unique())][['tr_id', 'gene', 'non_AUG_from_cds_seq']]
refseq_both.columns = ['tr_id', 'gene', 'start']
refseq_both.to_csv('tmp_res/refseq109_genes_with_both_AUG_and_non_AUG.txt', sep='\t', index=False)

# Gencode 35

metadata_pc_g35 = pd.read_csv('tmp_res/metadata_pc_g35.txt', sep='\t')

metadata_pc_g35 = metadata_pc_g35.drop('CDS_start', axis=1)

metadata_pc_g25 = pd.read_csv('tmp_res/metadata_pc_g25.txt', sep='\t')


df = read_gtf("data/gencode.v35.annotation.gtf")
df_genes = df[df["feature"] == "gene"]
df_CDS = df[df["feature"] == "CDS"]

annotation_CDS = df_CDS[['transcript_id', 'transcript_type', 'gene_name', 'tag']]
annotation_CDS.columns = ['tr_id', 'transcript_type', 'gene', 'tag']

df.columns

df[df['transcript_id'] == 'ENST00000594371.1'][['transcript_id', 'feature', 'transcript_type', 'tag']].drop_duplicates()

* non-AUG tag 

# we will create df with transcript type and tags 
li = []
                                                        
for i,feature in enumerate(DataIterator('data/gencode.v35.annotation.gtf')): 
    if i % 1000000 == 0:
        print (i)
    try:
        attr = feature.attributes
        gene_type = attr['gene_type'][0]
        transcript_id = attr['transcript_id'][0]
        gene_name = attr['gene_name'][0]
        tag = attr['tag']
        transcript_type = attr['transcript_type'][0]
            
        # cds_start_NF;cds_end_NF 
        if 'non_ATG_start' in tag:            
            li.append([transcript_id, gene_name, transcript_type, ';'.join(tag), 'non_ATG_start'])
        else:
            li.append([transcript_id, gene_name, transcript_type, ';'.join(tag), None])
            
    except Exception as e:
        continue           
    

df = pd.DataFrame(li, columns = ['tr_id', 'gene', 'biotype', 'tags', 'non-ATG-tag']).drop_duplicates()

# select genes with complete CDS, protein-coding transcript type and non-AUG tag 
non_AUG_annotated_pc_and_cds_complete = df[(df['non-ATG-tag'] == 'non_ATG_start') & (df['biotype'] == 'protein_coding') & 
                                                          (~df['tags'].str.contains('cds_start_NF')) &
                                                         (~df['tags'].str.contains('cds_end_NF'))]

non_AUG_annotated_pc_and_cds_complete.tr_id.nunique(), non_AUG_annotated_pc_and_cds_complete.gene.nunique()

non_AUG_annotated_pc_and_cds_complete[non_AUG_annotated_pc_and_cds_complete['gene'] == 'MYC'].to_csv('MYC.txt', sep='\t', index=False)

genes_with_non_AUG_proteoforms = list(non_AUG_annotated_pc_and_cds_complete['gene'].unique())

print (len(genes_with_non_AUG_proteoforms))

print ('TEAD1' in genes_with_non_AUG_proteoforms)

print ('TEAD3' in genes_with_non_AUG_proteoforms)

#genes_with_non_AUG_proteoforms

metadata_pc_g35[metadata_pc_g35['gene'] == 'TEAD1'][['tr_id',  'cds_start_codon', 'cds_stop_codon']].merge(annotation_CDS[annotation_CDS['gene'] == 'TEAD1'].drop_duplicates(), on='tr_id').to_csv('tmp_res/TEAD1.txt', sep='\t', index=False)

metadata_pc_g35[metadata_pc_g35['gene'] == 'MYC'][['tr_id',  'cds_start_codon', 'cds_stop_codon']].merge(annotation_CDS[annotation_CDS['gene'] == 'MYC'].drop_duplicates(), on='tr_id').to_csv('tmp_res/MYC.txt', sep='\t', index=False)

* different approach: get transcripts with full CDS, protein-coding type and extract start and stop as first 3 and last 3 nucleotides of CDS

cds_complete_and_pc_df = df[(df['biotype'] == 'protein_coding') & (~df['tags'].str.contains('cds_start_NF')) & (~df['tags'].str.contains('cds_end_NF'))]

print (cds_complete_and_pc_df.shape, df.shape)

t = metadata_pc_g35[['tr_id', 'cds_seq']].merge(cds_complete_and_pc_df, on=['tr_id'], how='inner')
print (t.shape, metadata_pc_g35.shape, )

li_ = []

for row in t.to_numpy():
    tr_id = row[0]
    gene = row[2]
    cds_seq = row[1]
    tags = row[-2]
    nonAUGtag = row[-1]
    
    start_CDS = cds_seq[0:3]
    stop_CDS = cds_seq[-3:]
    li_.append([tr_id, gene, tags, nonAUGtag, start_CDS, stop_CDS])
    
tmp = pd.DataFrame(li_, columns = ['tr_id', 'gene', 'tags', 'nonAUGtag', 'start_CDS', 'stop_CDS'])

* clearly something's wrong: starts and stops are weird. CDSs must be still truncated regardless of absent tags 

t = pd.DataFrame(tmp.groupby(by='start_CDS').count())[['tr_id']].reset_index().sort_values(by='tr_id', ascending=False)

t.columns = ['First_3nt_of_CDS', '#_transcripts']

t[t['First_3nt_of_CDS'] != 'ATG']['#_transcripts'].sum()

# truncated CDS even if there is no such tag in .gtf file 

tmp[(~tmp['gene'].isin(genes_with_non_AUG_proteoforms)) & (tmp['start_CDS'] != 'ATG')].to_csv('tmp_res/gencode35_troubled_genes_with_most_likely_truncated_CDS.txt', sep='\t', index=False)

tmp[(~tmp['gene'].isin(genes_with_non_AUG_proteoforms)) & (tmp['start_CDS'] != 'ATG')]

', '.join(list(tmp.stop_CDS.unique())), ', '.join(list(tmp.start_CDS.unique()))

tmp[tmp['start_CDS'] == 'TGA']

# KCNH7-205 ENST00000621889.1 has GGG as a start of translated sequence in gencode and it still has tags = basic...

# same with PACRG-210 ENST00000611387.1

we have to stick to non-AUG tag... 

# pc transcripts, complete cds -> merge with metadata
pr_trascript_cds_complete_df = cds_complete_and_pc_df.merge(metadata_pc_g35, on=['gene', 'tr_id'], how='inner')

# select genes with non-AUG
df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS = pr_trascript_cds_complete_df[pr_trascript_cds_complete_df['gene'].isin(genes_with_non_AUG_proteoforms)]

print (df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS.shape, df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS.gene.nunique(), Counter(df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS.cds_start_codon.tolist()))

df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS[df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS['cds_start_codon'] == 'AAC']

df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS = pr_trascript_cds_complete_df[pr_trascript_cds_complete_df['gene'].isin(genes_with_non_AUG_proteoforms)]

# CDS_start codon 
df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS['CDS_start'] = df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS['cds_seq'].str.slice(0, 3)

# gene - start codons dataframe
d1 = df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS.groupby('gene')['CDS_start'].apply(set).apply(list).apply(', '.join).to_dict()
gene_codon_g35_df = pd.DataFrame(zip(d1.keys(), d1.values()), columns = ['gene', 'codons'])

print ('there are both AUG and non-AUG proteoforms for %s genes' % gene_codon_g35_df[(gene_codon_g35_df['codons'].str.contains('ATG')) & (gene_codon_g35_df['codons'].str.len() >= 6)].shape[0])
print (', '.join(gene_codon_g35_df[(gene_codon_g35_df['codons'].str.contains('ATG')) & (gene_codon_g35_df['codons'].str.len() >= 6)].gene.unique()))
print ()

print ('there are only AUG codon proteoform for %s genes (although its non-AUG annotated)' % gene_codon_g35_df[(gene_codon_g35_df['codons'] == 'ATG')].shape[0])
print (', '.join(gene_codon_g35_df[(gene_codon_g35_df['codons'] == 'ATG')].gene.unique()))
print ()

print ('there are only non-AUG codon proteoform for %s genes' % gene_codon_g35_df[(gene_codon_g35_df['codons'].str.len() <=3) & (gene_codon_g35_df['codons'] != 'ATG')].shape[0])
print (', '.join(gene_codon_g35_df[(gene_codon_g35_df['codons'].str.len() <=3) & (gene_codon_g35_df['codons'] != 'ATG')].gene.unique()))

gene_codon_g35_df[(gene_codon_g35_df['codons'] == 'ATG')]

df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS[df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS['tr_id'] == 'ENST00000378179.9']

both_proteoforms_g35_genes = gene_codon_g35_df[(gene_codon_g35_df['codons'].str.contains('ATG')) & (gene_codon_g35_df['codons'].str.len() >= 6)].gene.unique()

only_non_AUG_proteoforms_genes = gene_codon_g35_df[(gene_codon_g35_df['codons'].str.len() <=3) & (gene_codon_g35_df['codons'] != 'ATG')].gene.unique()

print (len(only_non_AUG_proteoforms_genes), len(both_proteoforms_g35_genes))

both_proteoforms_g35_genes = gene_codon_g35_df[(gene_codon_g35_df['codons'].str.contains('ATG')) & (gene_codon_g35_df['codons'].str.len() >= 6)].gene.unique()

only_non_AUG_proteoforms_genes = gene_codon_g35_df[(gene_codon_g35_df['codons'].str.len() <=3) & (gene_codon_g35_df['codons'] != 'ATG')].gene.unique()

print (len(only_non_AUG_proteoforms_genes), len(both_proteoforms_g35_genes))

df = df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS.copy()

g35_only_df = df[df['gene'].isin(only_non_AUG_proteoforms_genes)][['tr_id', 'gene', 'cds_start_codon']]



g35_only_df2 = g35_only_df.merge(annotation_CDS, on=['tr_id', 'gene'], how='inner').drop_duplicates()

g35_only_df3 = g35_only_df2[(g35_only_df2['transcript_type'] == 'protein_coding') 
            & (~g35_only_df2['tag'].str.contains('CDS_start_NF')) &
            (~g35_only_df2['tag'].str.contains('CDS_end_NF'))]

print (g35_only_df.gene.nunique(), g35_only_df3.gene.nunique())


g35_only_df3.to_csv('tmp_res/gencode35_only_non_AUG_PC_complete_CDS.txt', sep='\t', index=False)

g35_both_df = df[df['gene'].isin(both_proteoforms_g35_genes)][['tr_id', 'gene', 'cds_start_codon']]

g35_both_df2 = g35_both_df.merge(annotation_CDS, on=['tr_id', 'gene'], how='inner').drop_duplicates()

g35_both_df3 = g35_both_df2[(g35_both_df2['transcript_type'] == 'protein_coding') 
            & (~g35_both_df2['tag'].str.contains('CDS_start_NF')) &
            (~g35_both_df2['tag'].str.contains('CDS_end_NF'))]

print (g35_both_df.gene.nunique(), g35_both_df3.gene.nunique())

g35_both_df3.to_csv('tmp_res/gencode35_both_AUG_non_AUG_PC_complete_CDS.txt', sep='\t', index=False)

* ENST00000467011.5 - no non_ATG_start tag in tags (g35), but transcript sequence starts with TTG
* there is ATG protein-coding transcript ENST00000478049.1 which has 'cds_end_NF' tag meaning truncated 3'end of CDS. Can it still be considered as protein-coding? 

# Comparison RefSeq and Gencode35

refseq_only_non_AUG['type_refseq'] = 'only'
refseq_only_non_AUG.columns = ['tr_id_refseq', 'gene', 'start', 'type_refseq']

refseq_both['type_refseq'] = 'both'
refseq_both.columns = ['tr_id_refseq', 'gene', 'start', 'type_refseq']

g35_only_df['type_g35'] = 'only'
g35_only_df.columns = ['tr_id_g35', 'gene', 'start', 'type_g35']

g35_both_df['type_g35'] = 'both'
g35_both_df.columns = ['tr_id_g35', 'gene', 'start', 'type_g35']

# only non-AUG both in g35 and refseq
print ('%s only non-AUG both in g35 and refseq' % len(list(set(refseq_only_non_AUG.gene.unique()).intersection(set(g35_only_df.gene.unique())))))
print (', '.join(list(set(refseq_only_non_AUG.gene.unique()).intersection(set(g35_only_df.gene.unique())))))

print ('%s genes non-AUG in g35' % g35_only_df.gene.nunique())
print ('%s genes non-AUG in refseq' % refseq_only_non_AUG.gene.nunique())

set(g35_only_df.gene.unique()) - set(refseq_only_non_AUG.gene.unique()).intersection(set(g35_only_df.gene.unique()))

', '.join(list(set(refseq_only_non_AUG.gene.unique()) - set(refseq_only_non_AUG.gene.unique()).intersection(set(g35_only_df.gene.unique()))))

g35_only_df[g35_only_df['gene'] == 'TUG1']

refseq_only_non_AUG[refseq_only_non_AUG['gene'] == 'TUG1']

----

g35_both_df

# only non-AUG both in g35 and refseq
print ('%s both AUG and non-AUG both in g35 and refseq' % len(list(set(refseq_both.gene.unique()).intersection(set(g35_both_df.gene.unique())))))
print (', '.join(list(set(refseq_both.gene.unique()).intersection(set(g35_both_df.gene.unique())))))

print ('%s genes non-AUG in g35' % g35_both_df.gene.nunique())
print ('%s genes non-AUG in refseq' % refseq_both.gene.nunique())

', '.join(list(set(g35_both_df.gene.unique()) - set(refseq_both.gene.unique()).intersection(set(g35_both_df.gene.unique()))))

', '.join(list(set(refseq_both.gene.unique()) - set(refseq_both.gene.unique()).intersection(set(g35_both_df.gene.unique()))))

# Sanity check: PhyloSET and RiboSET overlaps 

ps = pd.read_csv('tmp_res/SET1.txt', sep='\t')
rb =  pd.read_csv('tmp_res/SET2.txt', sep='\t')

set(ps.gene.unique()).intersection(g35_both_df.gene.tolist()), set(ps.gene.unique()).intersection(g35_only_df.gene.tolist())

# OAZ3: extensions is different (diff transcripts, not overlapping, diff 5'UTRs, just gene name match)
# CACNG8: same transcript, CTG-init, ext is found in 5'UTR before annotated CTG (positive score)

set(ps.gene.unique()).intersection(refseq_both.gene.tolist()), set(ps.gene.unique()).intersection(refseq_only_non_AUG.gene.tolist())

set(rb.gene.unique()).intersection(g35_both_df.gene.tolist()), set(rb.gene.unique()).intersection(g35_only_df.gene.tolist())
# STIM2: same transcript, CTG-init, ext is found in 5'UTR before annotated CTG (positive score)

set(rb.gene.unique()).intersection(refseq_both.gene.tolist()), set(rb.gene.unique()).intersection(refseq_only_non_AUG.gene.tolist())

-----

# Compare gencode25/gencode35

metadata_pc_g25 = pd.read_csv('tmp_res/metadata_pc_g25.txt', sep='\t')


df_g25 = read_gtf("data/gencode.v25.annotation.gtf")
df_genes_g25 = df_g25[df_g25["feature"] == "gene"]
df_CDS_g25 = df_g25[df_g25["feature"] == "CDS"]

annotation_CDS_g25 = df_CDS_g25[['transcript_id', 'transcript_type', 'gene_name', 'tag']]
annotation_CDS_g25.columns = ['tr_id', 'transcript_type', 'gene', 'tag']

### Non-AUG tag

# we will create df with transcript type and tags 
li = []
                                                        
for i,feature in enumerate(DataIterator('data/gencode.v25.annotation.gtf')): 
    if i % 1000000 == 0:
        print (i)
    try:
        attr = feature.attributes
        gene_type = attr['gene_type'][0]
        transcript_id = attr['transcript_id'][0]
        gene_name = attr['gene_name'][0]
        tag = attr['tag']
        transcript_type = attr['transcript_type'][0]
            
        # cds_start_NF;cds_end_NF 
        if 'non_ATG_start' in tag:            
            li.append([transcript_id, gene_name, transcript_type, ';'.join(tag), 'non_ATG_start'])
        else:
            li.append([transcript_id, gene_name, transcript_type, ';'.join(tag), None])
            
    except Exception as e:
        continue           
    

df = pd.DataFrame(li, columns = ['tr_id', 'gene', 'biotype', 'tags', 'non-ATG-tag']).drop_duplicates()

# select genes with complete CDS, protein-coding transcript type and non-AUG tag 
non_AUG_annotated_pc_and_cds_complete_g25 = df[(df['non-ATG-tag'] == 'non_ATG_start') & (df['biotype'] == 'protein_coding') & 
                                                          (~df['tags'].str.contains('cds_start_NF')) &
                                                         (~df['tags'].str.contains('cds_end_NF'))]

genes_with_non_AUG_proteoforms_g25 = list(non_AUG_annotated_pc_and_cds_complete_g25['gene'].unique())

print (len(genes_with_non_AUG_proteoforms_g25))

cds_complete_and_pc_df_g25 = df[(df['biotype'] == 'protein_coding') & (~df['tags'].str.contains('cds_start_NF')) & (~df['tags'].str.contains('cds_end_NF'))]

# pc transcripts, complete cds -> merge with metadata
pr_trascript_cds_complete_df_g25 = cds_complete_and_pc_df_g25.merge(metadata_pc_g25, on=['gene', 'tr_id'], how='inner')

# select genes with non-AUG
df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS_g25 = pr_trascript_cds_complete_df_g25[pr_trascript_cds_complete_df_g25['gene'].isin(genes_with_non_AUG_proteoforms_g25)]

print (df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS_g25.shape, df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS_g25.gene.nunique(), Counter(df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS_g25.cds_start_codon.tolist()))

# CDS_start codon 
df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS_g25['CDS_start'] = df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS_g25['cds_seq'].str.slice(0, 3)

# gene - start codons dataframe
d1 = df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS_g25.groupby('gene')['CDS_start'].apply(set).apply(list).apply(', '.join).to_dict()
gene_codon_g25_df = pd.DataFrame(zip(d1.keys(), d1.values()), columns = ['gene', 'codons'])

print ('there are both AUG and non-AUG proteoforms for %s genes' % gene_codon_g25_df[(gene_codon_g25_df['codons'].str.contains('ATG')) & (gene_codon_g25_df['codons'].str.len() >= 6)].shape[0])
print (', '.join(gene_codon_g25_df[(gene_codon_g25_df['codons'].str.contains('ATG')) & (gene_codon_g25_df['codons'].str.len() >= 6)].gene.unique()))
print ()

print ('there are only AUG codon proteoform for %s genes (although its non-AUG annotated)' % gene_codon_g35_df[(gene_codon_g35_df['codons'] == 'ATG')].shape[0])
print (', '.join(gene_codon_g25_df[(gene_codon_g25_df['codons'] == 'ATG')].gene.unique()))
print ()

print ('there are only non-AUG codon proteoform for %s genes' % gene_codon_g25_df[(gene_codon_g25_df['codons'].str.len() <=3) & (gene_codon_g25_df['codons'] != 'ATG')].shape[0])
print (', '.join(gene_codon_g25_df[(gene_codon_g25_df['codons'].str.len() <=3) & (gene_codon_g25_df['codons'] != 'ATG')].gene.unique()))

### both AUG/nonAUG or only nonAUG dfs

both_proteoforms_g25_genes = gene_codon_g25_df[(gene_codon_g25_df['codons'].str.contains('ATG')) & (gene_codon_g25_df['codons'].str.len() >= 6)].gene.unique()

only_non_AUG_proteoforms_genes_g25 = gene_codon_g25_df[(gene_codon_g25_df['codons'].str.len() <=3) & (gene_codon_g25_df['codons'] != 'ATG')].gene.unique()

print (len(only_non_AUG_proteoforms_genes_g25), len(both_proteoforms_g25_genes))

df = df_with_all_proteoforms_of_genes_with_non_AUG_complete_CDS_g25.copy()

g25_only_df = df[df['gene'].isin(only_non_AUG_proteoforms_genes_g25)][['tr_id', 'gene', 'cds_start_codon']]



g25_only_df2 = g25_only_df.merge(annotation_CDS_g25, on=['tr_id', 'gene'], how='inner').drop_duplicates()

g25_only_df3 = g25_only_df2[(g25_only_df2['transcript_type'] == 'protein_coding') 
            & (~g25_only_df2['tag'].str.contains('CDS_start_NF')) &
            (~g25_only_df2['tag'].str.contains('CDS_end_NF'))]

print (g25_only_df.gene.nunique(), g25_only_df3.gene.nunique())


g25_only_df3.to_csv('tmp_res/gencode25_only_non_AUG_PC_complete_CDS.txt', sep='\t', index=False)

g25_both_df = df[df['gene'].isin(both_proteoforms_g25_genes)][['tr_id', 'gene', 'cds_start_codon']]

g25_both_df2 = g25_both_df.merge(annotation_CDS_g25, on=['tr_id', 'gene'], how='inner').drop_duplicates()

g25_both_df3 = g25_both_df2[(g25_both_df2['transcript_type'] == 'protein_coding') 
            & (~g25_both_df2['tag'].str.contains('CDS_start_NF')) &
            (~g25_both_df2['tag'].str.contains('CDS_end_NF'))]

print (g25_both_df.gene.nunique(), g25_both_df3.gene.nunique())

g25_both_df3.to_csv('tmp_res/gencode25_both_AUG_non_AUG_PC_complete_CDS.txt', sep='\t', index=False)

---------------

g25_only_df['type_g25'] = 'only'
g25_only_df.columns = ['tr_id_g25', 'gene', 'start', 'type_g25']

g25_both_df['type_g25'] = 'both'
g25_both_df.columns = ['tr_id_g25', 'gene', 'start', 'type_g25']

# only non-AUG both in g35 and g25
print ('%s only non-AUG both in g25 and refseq' % len(list(set(g35_only_df.gene.unique()).intersection(set(g25_only_df.gene.unique())))))
print (', '.join(list(set(g35_only_df.gene.unique()).intersection(set(g25_only_df.gene.unique())))))

print ('%s genes non-AUG in g25' % g25_only_df.gene.nunique())
print ('%s genes non-AUG in g35' % g35_only_df.gene.nunique())

# only non-AUG both in g35 and g25
print ('%s both AUG and non-AUG both in g35 and refseq' % len(list(set(g25_both_df.gene.unique()).intersection(set(g35_both_df.gene.unique())))))
print (', '.join(list(set(g25_both_df.gene.unique()).intersection(set(g35_both_df.gene.unique())))))

print ('%s genes non-AUG in g25' % g25_both_df.gene.nunique())
print ('%s genes non-AUG in g35' % g35_both_df.gene.nunique())
