==> jupyter_notebooks/11_gene_enrichment_analysys.ipynb <==
%matplotlib inline

import numpy as np
import collections
from collections import OrderedDict, Counter, defaultdict
import pandas as pd

import Bio
from Bio import SeqIO

import seaborn as sns
import matplotlib.pyplot as plt
from bokeh.plotting import figure, output_file, show 
import matplotlib.patches as patches    
from matplotlib import colors


import glob

import subprocess
from subprocess import call

import re

import pickle

from Bio.SubsMat import MatrixInfo
import Bio.Data.CodonTable

from scipy import stats


from Bio.SubsMat import MatrixInfo
import Bio.Data.CodonTable

# open metadata with scores and metrics
meta = pd.read_csv('tmp_res/METADATA_CURRENT.txt', sep='\t')

# gencode 25
metadata_pc_g25 = pd.read_csv('tmp_res/metadata_pc_g25.txt', sep='\t')


P_Set = pd.read_csv('tmp_res/SET1.txt', sep='\t')

R_Set = pd.read_csv('tmp_res/SET2.txt', sep='\t')



R_Set['pos_of_codon'] = [int(x.split('; ')[1].split('-')[0])-1 for x in R_Set.Riboseq_Summary.tolist()]

R_Set['codon'] = [x.split('; ')[0] for x in R_Set.Riboseq_Summary.tolist()]

R_Set = R_Set[(~R_Set['tr_id'].str.contains('PAR')) & (R_Set['codon'] != 'ATG')]

R_Set[0:2]

R_Set[R_Set['gene'] == 'CTD-2135J3.4']

# GSEA 

%matplotlib inline
%config InlineBackend.figure_format='retina' # mac
%load_ext autoreload
%autoreload 2
import pandas as pd
import gseapy as gp
import matplotlib.pyplot as plt

print (gp.__version__)

names = gp.get_library_name(database='Human') # default: Human
names[:10]

len(names)

[x if x.find('KEGG') != -1 else None for x in names]

# EnrichR

### RiboSET 

gene_list_rset = list(R_Set.gene.unique())

enrichR_set = {}

for i, gene_set in enumerate(names):
    
    enr = gp.enrichr(gene_list=gene_list_rset,
                 gene_sets=[gene_set],
                 organism='Human', # don't forget to set organism to the one you desired! e.g. Yeast
                 description='test_name',
                 outdir='data/GeneEnrichment/enrichr_kegg',
                 # no_plot=True,
                 cutoff=0.1 # test dataset, use lower value from range(0,1)
                )
    # select only significant 
    enr_res_sel = enr.results[enr.results['Adjusted P-value'] <= 0.05]
    enrichR_set[gene_set] = enr_res_sel
    
    print (i, gene_set, '# of sets with adj pv < 0.05', enr_res_sel.shape[0])

Aging_Perturbations_from_GEO_down, 
Aging_Perturbations_from_GEO_up
Allen_Brain_Atlas_10x_scRNA_2021 
COVID-19_Related_Gene_Sets_2021
DSigDB 
Disease_Perturbations_from_GEO_down 
Disease_Perturbations_from_GEO_up 
Drug_Perturbations_from_GEO_2014 
GeneSigDB

MSigDB_Oncogenic_Signatures 
NCI-60_Cancer_Cell_Lines  
Rare_Diseases_AutoRIF_Gene_Lists 
''''''
Genes_Associated_with_NIH_Grants 
''''''

Human_Gene_Atlas
Jensen_COMPARTMENTS 
InterPro_Domains_2019 
Pfam_Domains_2019

ProteomicsDB_2020 
SILAC_Phosphoproteomics

SysMyo_Muscle_Gene_Sets

TargetScan_microRNA_2017


Transcription_Factor_PPIs

Virus-Host_PPI_P-HIPSTer_2020

lncHUB_lncRNA_Co-Expression

pd.concat([enrichR_set['Aging_Perturbations_from_GEO_up'], enrichR_set['Aging_Perturbations_from_GEO_down']]).to_csv('Supplementary_tables_and_plots/tables/Gene_enrichment_RiboSET_aging.txt', sep='\t', index=False)

enrichR_set['COVID-19_Related_Gene_Sets_2021'].to_csv('Supplementary_tables_and_plots/tables/Gene_enrichment_RiboSET_covid19.txt', sep='\t', index=False)

enrichR_set['Rare_Diseases_AutoRIF_Gene_Lists'].to_csv('Supplementary_tables_and_plots/tables/Rare_Diseases_AutoRIF_Gene_Lists_Riboseq.txt', sep='\t', index=False)

pd.concat([enrichR_set['Disease_Perturbations_from_GEO_down'], enrichR_set['Disease_Perturbations_from_GEO_down']]).to_csv('Supplementary_tables_and_plots/tables/Gene_enrichment_RiboSET_Disease_Perturbations_riboseq.txt', sep='\t', index=False)

there were 2 gene enriment hits with terms in Rare_Diseases_AutoRIF_Gene_Lists: Periventricular leukomalacia and Leukomalacia. 

### PhyloSET

gene_list_pset = list(P_Set.gene.unique())
len(gene_list_pset)

enrichR_set_phyloset = {}

for i, gene_set in enumerate(names):
    
    enr = gp.enrichr(gene_list=gene_list_pset,
                 gene_sets=[gene_set],
                 organism='Human', # don't forget to set organism to the one you desired! e.g. Yeast
                 description='test_name',
                 outdir='data/GeneEnrichment/enrichr_kegg',
                 # no_plot=True,
                 cutoff=0.1 # test dataset, use lower value from range(0,1)
                )
    # select only significant 
    enr_res_sel = enr.results[enr.results['Adjusted P-value'] <= 0.05]
    enrichR_set_phyloset[gene_set] = enr_res_sel
    
    print (i, gene_set, '# of sets with adj pv < 0.05', enr_res_sel.shape[0])

enrichR_set_phyloset['RNA-Seq_Disease_Gene_and_Drug_Signatures_from_GEO'].to_csv('Supplementary_tables_and_plots/tables/Gene_enrichment_PhyloSET_RNA.txt', sep='\t', index=False)

enrichR_set_phyloset['RNA-Seq_Disease_Gene_and_Drug_Signatures_from_GEO']

enrichR_set_phyloset['MGI_Mammalian_Phenotype_Level_4_2021']

enrichR_set_phyloset['Jensen_DISEASES']

enrichR_set_phyloset['Jensen_TISSUES'].to_csv('Supplementary_tables_and_plots/tables/Gene_enrichment_PhyloSET_Jensen_TISSUES.txt', sep='\t', index=False)

--------

## Plot enrichR 

sns.scatterplot(data=enr_res_sel, x="Adjusted P-value", y="Combined Score")

tmp = enrichR_set['COVID-19_Related_Gene_Sets_2021'][['Term', 'Adjusted P-value', 'Combined Score', 'Genes', 'Overlap']]
tmp['hits'] = [len(x.split(';')) for x in tmp['Genes']]
tmp['-log.adj.pv'] = [-1*np.log10(x) for x in tmp['Adjusted P-value'].tolist()]

tmp.head(10) 

sns.set(font_scale=3)  # crazy big

sns.set_style("white")

fig, ax = plt.subplots(figsize=(10, 5))

ax = sns.scatterplot(x=tmp.head(10)['-log.adj.pv'], 
              y=tmp.head(10).Term, 
             size=tmp.head(10).hits*4,
             hue = tmp.head(10)['Combined Score'])

plt.legend(bbox_to_anchor=(1.05, 1), loc=2, borderaxespad=0.)

#handles, labels = ax.get_legend_handles_labels()

#ax._legend.remove()
#ax.legend(handles, labels, ncol=2, loc='upper center', 
#                bbox_to_anchor=(1.4, 1), frameon=False)

------

enrichR_set['Jensen_COMPARTMENTS']['Term'].tolist()

enrichR_set['InterPro_Domains_2019']

Thioredoxin domain (7/40 overlap, adj.p-value = 0.002582), Ubiquitin-conjugating enzyme E2 (6/40 overlap, adj.p-value = 0.000121)

enrichR_set['Pfam_Domains_2019']

enrichR_set['DSigDB']

enrichR_set['Rare_Diseases_AutoRIF_Gene_Lists']

Aging_Perturbations_from_GEO_down, 
Aging_Perturbations_from_GEO_up
Allen_Brain_Atlas_10x_scRNA_2021 
COVID-19_Related_Gene_Sets_2021
DSigDB 
Disease_Perturbations_from_GEO_down 
Disease_Perturbations_from_GEO_up 
Drug_Perturbations_from_GEO_2014 
GeneSigDB

MSigDB_Oncogenic_Signatures 
NCI-60_Cancer_Cell_Lines  
Rare_Diseases_AutoRIF_Gene_Lists 
''''''
Genes_Associated_with_NIH_Grants 
''''''

Human_Gene_Atlas
Jensen_COMPARTMENTS 
InterPro_Domains_2019 
Pfam_Domains_2019

ProteomicsDB_2020 
SILAC_Phosphoproteomics

SysMyo_Muscle_Gene_Sets

TargetScan_microRNA_2017


Transcription_Factor_PPIs

Virus-Host_PPI_P-HIPSTer_2020

lncHUB_lncRNA_Co-Expression

# GO terms 

# ontology basic file 
# can be retireved by:
# from goatools.base import download_go_basic_obo
# obo_fname = download_go_basic_obo()

# download associations??

# Get ftp://ftp.ncbi.nlm.nih.gov/gene/DATA/gene2go.gz
#from goatools.base import download_ncbi_associations
#fin_gene2go = download_ncbi_associations()

## Load files 

from goatools.obo_parser import GODag

obodag = GODag("data/GeneEnrichment/go-basic.obo")

from __future__ import print_function
from goatools.base import download_ncbi_associations
from goatools.base import download_go_basic_obo
from goatools.anno.genetogo_reader import Gene2GoReader

# Read NCBI's gene2go. Store annotations in a list of namedtuples
objanno = Gene2GoReader('data/GeneEnrichment/gene2go', taxids=[9606])

# Get namespace2association where:
#    namespace is:
#        BP: biological_process               
#        MF: molecular_function
#        CC: cellular_component
#    assocation is a dict:
#        key: NCBI GeneID
#        value: A set of GO IDs associated with that gene
ns2assoc = objanno.get_ns2assc()

for nspc, id2gos in ns2assoc.items():
    print("{NS} {N:,} annotated mouse genes".format(NS=nspc, N=len(id2gos)))

import goatools

print(goatools.__version__)

# pip install goatools



### Load Background gene set 

#### Open HGNC gene conversion table

ids_df1 = pd.read_csv('data/GeneEnrichment/HGNC_id_to_NCBI_geneID.txt', sep='\t')

# ids_df1 = HGNC gene conversion table
# entrez_gene_ids_df
# ids_df = gene info

#### Open entrez ID metadata table 

entrez_gene_ids_df = pd.read_csv('data/GeneEnrichment/gencode.v38.metadata.EntrezGene', sep='\t', header=None)
entrez_gene_ids_df.columns = ['tr_id', 'entrez_id']

print (entrez_gene_ids_df.merge(metadata_pc_g25[['tr_id', 'gene']], on=['tr_id']).gene.nunique())

#### Open gene_info table from ncbi 

# Such info is stored in https://ftp.ncbi.nlm.nih.gov/gene/DATA/gene_info.gz
# take all human genes 

ids_df = pd.read_csv('data/GeneEnrichment/gene_info', sep='\t', 
                     usecols=['#tax_id', 'GeneID','Symbol'], low_memory=True)

ids_df_sel = ids_df[(ids_df['Symbol'] != 'NEWENTRY') & (ids_df['#tax_id'] == 9606)]

print (ids_df_sel.Symbol.nunique(), ids_df_sel.GeneID.nunique())

------

#### Transform gene symbols from gencode25 to entrez gene id (with some missing genes)

ids_df_sel_g25 = ids_df_sel[ids_df_sel['Symbol'].isin(metadata_pc_g25.gene.tolist())].drop_duplicates()

ids_df_sel_g25_LI = ids_df_sel_g25.GeneID.tolist()

remaining = list(set(metadata_pc_g25.gene.tolist()) - set(ids_df_sel_g25.Symbol.tolist()))
print (len(remaining))

ids_df3 = ids_df1[ids_df1['Previous symbols'].notna()]

prev_ids = pd.concat([pd.Series(row['NCBI Gene ID(supplied by NCBI)'], row['Previous symbols'].split(','))              
                    for _, row in ids_df3.iterrows()]).reset_index()

add_548_missing_df = prev_ids[prev_ids['index'].isin(remaining)]
add_548_missing_LI = prev_ids[prev_ids['index'].isin(remaining)][0].tolist()

remaining2 = set(remaining) -  set(prev_ids[prev_ids['index'].isin(remaining)]['index'].tolist())
print (len(remaining2))

ids_df2 = ids_df1[ids_df1['Alias symbols'].notna()]

alias_ids = pd.concat([pd.Series(row['NCBI Gene ID(supplied by NCBI)'], row['Alias symbols'].split(','))              
                    for _, row in ids_df2.iterrows()]).reset_index()

add_10_missing2_LI = alias_ids[alias_ids['index'].isin(remaining2)][0].tolist()

remaining3 = set(remaining2) - set(alias_ids[alias_ids['index'].isin(remaining2)]['index'].tolist())

save_remaining3_df = metadata_pc_g25[metadata_pc_g25['gene'].isin(remaining3)].merge(entrez_gene_ids_df, on='tr_id')[['entrez_id', 'gene', 'tr_id']]

save_remaining3_LI = save_remaining3_df.entrez_id.tolist()

print (len(save_remaining3_LI))

remaining4 =  set(remaining3) - set(save_remaining3_df.gene.tolist())

print (len(remaining4))

---------

total_found = list(set(ids_df_sel_g25_LI + add_548_missing_LI + add_10_missing2_LI + save_remaining3_LI))
print (len(total_found))
print (metadata_pc_g25.gene.nunique())

20330-19621

### transform gene symbols intp entrez gene ids

ids_df_sel[0:2]

Rset_entrez_ids = ids_df_sel[ids_df_sel['Symbol'].isin(R_Set['gene'].tolist())].GeneID.tolist()

list(set(R_Set.gene.tolist()) -  set(ids_df_sel[ids_df_sel['Symbol'].isin(R_Set['gene'].tolist())].Symbol.tolist()))

# ids_df1[ids_df1['Previous symbols'].isin(list(set(R_Set.gene.tolist()) -  set(Rset_entrez_ids.Symbol.tolist())))]

d = {'NARS':4677,
 'CTD-2135J3.4':7013, #PIN2 TERF
 'CTGF':1490,
 'C18orf8':29919, 
 'HN1':51155,
 'LARS':51520,
 'MGEA5':10724,
 'KIF1BP':26128, 
 'C9orf16': 79095,
 'FAM64A':54478,
 'RRNAD1': 51093,
 'YARS':8565,
 'AC013461.1':51776, #ZAK
 'H1FX':8971,
 'AARS':16,
 'FAM92A1':137392, 
 'FAM160B1':57700,  
 'EPRS':2058,   
 'TMEM189':387521,
 'UFD1L':7353}

ids_df2[ids_df2['Alias symbols'].str.contains('PIN2')]

ids_df3 = ids_df1[ids_df1['Previous symbols'].notna()]

ids_df3[ids_df3['Previous symbols'].str.contains('AC013461.1')]

Rset_entrez_ids_full = Rset_entrez_ids + list(d.values())

len(Rset_entrez_ids), len(Rset_entrez_ids_full)

rset_convert_dict = dict(zip(ids_df_sel[ids_df_sel['Symbol'].isin(R_Set['gene'].tolist())].GeneID.tolist(), ids_df_sel[ids_df_sel['Symbol'].isin(R_Set['gene'].tolist())].Symbol.tolist()))


d1 = {v: k for k, v in d.items()}

rset_convert_dict.update(d1)

print (len(list(rset_convert_dict.keys())))

### Initialize a GOEA object


GeneID_human = total_found # our gencode25 gene database of converted entrez GeneID
print (len(GeneID_human))

from goatools.goea.go_enrichment_ns import GOEnrichmentStudyNS
import fisher

goeaobj = GOEnrichmentStudyNS(
        GeneID_human, # List of human protein-coding genes
        ns2assoc, # geneid/GO associations
        obodag, # Ontologies
        propagate_counts = False,
        alpha = 0.1, # default significance cut-off
        methods = ['fdr_bh']) # defult multipletest correction method

### Run Gene Ontology Enrichment Analysis 

# 'p_' means "pvalue". 'fdr_bh' is the multipletest method we are currently using.
geneids_study = Rset_entrez_ids_full

goea_results_all = goeaobj.run_study(geneids_study)

goea_results_sig = [r for r in goea_results_all if r.p_fdr_bh < 0.05]

### significant results 

print('{N} of {M:,} results were significant'.format(
    N=len(goea_results_sig),
    M=len(goea_results_all)))

### enriched vs purified results 

print('Significant results: {E} enriched, {P} purified'.format(
    E=sum(1 for r in goea_results_sig if r.enrichment=='e'),
    P=sum(1 for r in goea_results_sig if r.enrichment=='p')))

### Significant GOEA results by namespace

import collections as cx
ctr = cx.Counter([r.NS for r in goea_results_sig])
print('Significant results[{TOTAL}] = {BP} BP + {MF} MF + {CC} CC'.format(
    TOTAL=len(goea_results_sig),
    BP=ctr['BP'],  # biological_process
    MF=ctr['MF'],  # molecular_function
    CC=ctr['CC'])) # cellular_component

### Plot significant GO terms (likely to be messy)

from goatools.godag_plot import plot_gos, plot_results, plot_goid2goobj

plot_results("nbt3102_{NS}.png", goea_results_sig)

# Transform significant results into table

li = []

for el in goea_results_sig:
    go_term  = el.GO
    meaning = el.name
    hits_sample = '/'.join([str(x) for x in list(el.ratio_in_pop)])
    hits_bg = '/'.join([str(x) for x in list(el.ratio_in_study)])
    pv = el.p_fdr_bh
    genes_ids = list(el.study_items)  
    gene_symbols = list(map(rset_convert_dict.get, genes_ids))
    
    if el.NS == 'BP':
        ns = 'biological_process'
    elif el.NS == 'MF':
        ns = 'molecular_function'
    elif el.NS == 'CC':
        ns = 'cellular_component'
        
    level = el.goterm.level
    
    
    li.append([go_term, ns, meaning, level, hits_sample, hits_bg, pv, ', '.join([str(x) for x in genes_ids]), 
               ', '.join([str(x) for x in gene_symbols])])

rset_go_significant = pd.DataFrame(li, columns = ['GO_term', 'category', 'name', 'level',  'hits_sample', 'hits_background', 'pv_adj_fdr', 'NCBIgeneIDs', 'gene_symbols'])

rset_go_significant.to_csv('Supplementary_tables_and_plots/tables/rset_GO_significant.txt', sep='\t', index=False)

nucleoplasm = rset_go_significant[rset_go_significant['name'] == 'nucleoplasm'].iloc[0].gene_symbols.split(', ')
cytoplasm = rset_go_significant[rset_go_significant['name'] == 'cytoplasm'].iloc[0].gene_symbols.split(', ')
membrane = rset_go_significant[rset_go_significant['name'] == 'membrane'].iloc[0].gene_symbols.split(', ')

print (len(nucleoplasm), len(cytoplasm), len(membrane))
print (len(set(cytoplasm).intersection(nucleoplasm)))
print (len(set(cytoplasm).intersection(membrane)))
print (len(set(nucleoplasm).intersection(membrane)))

print (len(set(nucleoplasm).intersection(membrane).intersection(cytoplasm)))

132-7-9-49, 69-7-20-9, 139-9-49-20

rna_binding = rset_go_significant[rset_go_significant['name'] == 'RNA binding'].iloc[0].gene_symbols.split(', ')
protein_binding = rset_go_significant[rset_go_significant['name'] == 'protein binding'].iloc[0].gene_symbols.split(', ')

print (len(rna_binding), len(protein_binding))
print (len(set(protein_binding).intersection(rna_binding)))

In addition, significantly enriched 'RNA binding' term (59 genes) and 'protein binding' term (323 genes) have 51 common genes

Interestingly, 58 commons genes were found in enriched gene sets associated with term 'nucleoplasm' (132 genes) and 'cytoplasm' (139 genes). 

rset_go_significant

# PhyloSET 

### transform gene symbols into gene IDs 

# ids_df1 = HGNC gene conversion table
# entrez_gene_ids_df
# ids_df = gene info

ids_df2 = ids_df1[ids_df1['Previous symbols'].notna()]
ids_df3 = ids_df1[ids_df1['Alias symbols'].notna()]

pset_gene_conversion = ids_df[(ids_df['Symbol'].isin(P_Set.gene.tolist())) & (ids_df['#tax_id'] == 9606)][['GeneID', 'Symbol']].drop_duplicates()

remaining  = list(set(P_Set.gene.tolist()) - set(pset_gene_conversion.Symbol.tolist()))
remaining

d = {8971: 'H1FX', 
    # '': 'AC007906.1',  # RP11-295M3 # ENST00000624244.1 (AC007906.1)
    554250: 'GDF5OS', 
    26083: 'TBC1D29'}
d

pset_gene_conversion_dict = dict(zip(pset_gene_conversion.GeneID.tolist(), pset_gene_conversion.Symbol.tolist()))

pset_gene_conversion_dict.update(d)

pset_gene_conversion_li = list(pset_gene_conversion_dict.keys())

print (len(list(pset_gene_conversion_dict.keys())))

ids_df2[ids_df2['Previous symbols'].str.contains('GC16P053202')] # Previous symbols  Alias symbols 

ids_df3[ids_df3['Alias symbols'].str.contains('RP11-295M3')] # Previous symbols  Alias symbols 

ids_df1[ids_df1['Approved symbol'] == 'RP11-295M3'] # Previous symbols  Alias symbols 

ids_df[ids_df['Symbol'] == 'RP11-295M3']

P_Set[P_Set['gene'] == 'AC007906.1']

entrez_gene_ids_df[entrez_gene_ids_df['tr_id'] == 'ENST00000624244.1']

### GO enrichment for PhyloSET

# 'p_' means "pvalue". 'fdr_bh' is the multipletest method we are currently using.
geneids_study = pset_gene_conversion_li

goea_results_all = goeaobj.run_study(geneids_study)

goea_results_sig = [r for r in goea_results_all if r.p_fdr_bh < 0.05]

print('{N} of {M:,} results were significant'.format(
    N=len(goea_results_sig),
    M=len(goea_results_all)))

print('Significant results: {E} enriched, {P} purified'.format(
    E=sum(1 for r in goea_results_sig if r.enrichment=='e'),
    P=sum(1 for r in goea_results_sig if r.enrichment=='p')))

import collections as cx
ctr = cx.Counter([r.NS for r in goea_results_sig])
print('Significant results[{TOTAL}] = {BP} BP + {MF} MF + {CC} CC'.format(
    TOTAL=len(goea_results_sig),
    BP=ctr['BP'],  # biological_process
    MF=ctr['MF'],  # molecular_function
    CC=ctr['CC'])) # cellular_component

from goatools.godag_plot import plot_gos, plot_results, plot_goid2goobj

plot_results("Pset_nbt3102_{NS}.png", goea_results_sig)

goea_results_sig

# GO enrichment for RiboSET_ext

R_Set_ext = pd.read_csv('revision/Supplementary_tables/Supplementary_table_2_RiboSET_ext.txt', sep='\t')

R_Set_ext[0:4]

pset_gene_conversion = ids_df[(ids_df['Symbol'].isin(R_Set_ext.gene.tolist())) & (ids_df['#tax_id'] == 9606)][['GeneID', 'Symbol']].drop_duplicates()

remaining  = list(set(P_Set.gene.tolist()) - set(pset_gene_conversion.Symbol.tolist()))
remaining

len(pset_gene_conversion)

d = {8971: 'H1FX', 
    # '': 'AC007906.1',  # RP11-295M3 # ENST00000624244.1 (AC007906.1)
    554250: 'GDF5OS', 
    26083: 'TBC1D29'}
d

pset_gene_conversion_dict = dict(zip(pset_gene_conversion.GeneID.tolist(), pset_gene_conversion.Symbol.tolist()))

#pset_gene_conversion_dict.update(d)

pset_gene_conversion_li = list(pset_gene_conversion_dict.keys())

print (len(list(pset_gene_conversion_dict.keys())))

# 'p_' means "pvalue". 'fdr_bh' is the multipletest method we are currently using.
geneids_study = pset_gene_conversion_li

goea_results_all = goeaobj.run_study(geneids_study)

goea_results_sig = [r for r in goea_results_all if r.p_fdr_bh < 0.05]

print('{N} of {M:,} results were significant'.format(
    N=len(goea_results_sig),
    M=len(goea_results_all)))

print('Significant results: {E} enriched, {P} purified'.format(
    E=sum(1 for r in goea_results_sig if r.enrichment=='e'),
    P=sum(1 for r in goea_results_sig if r.enrichment=='p')))

import collections as cx
ctr = cx.Counter([r.NS for r in goea_results_sig])
print('Significant results[{TOTAL}] = {BP} BP + {MF} MF + {CC} CC'.format(
    TOTAL=len(goea_results_sig),
    BP=ctr['BP'],  # biological_process
    MF=ctr['MF'],  # molecular_function
    CC=ctr['CC'])) # cellular_component

from goatools.godag_plot import plot_gos, plot_results, plot_goid2goobj

plot_results("Pset_nbt3102_{NS}.png", goea_results_sig)

li = []

for el in goea_results_sig:
    go_term  = el.GO
    meaning = el.name
    hits_sample = '/'.join([str(x) for x in list(el.ratio_in_pop)])
    hits_bg = '/'.join([str(x) for x in list(el.ratio_in_study)])
    pv = el.p_fdr_bh
    genes_ids = list(el.study_items)  

    gene_symbols = list(map(pset_gene_conversion_dict.get, genes_ids))
    
    if el.NS == 'BP':
        ns = 'biological_process'
    elif el.NS == 'MF':
        ns = 'molecular_function'
    elif el.NS == 'CC':
        ns = 'cellular_component'
        
    level = el.goterm.level
    
    
    li.append([go_term, ns, meaning, level, hits_sample, hits_bg, pv, ', '.join([str(x) for x in genes_ids]), 
               ', '.join([str(x) for x in gene_symbols])])

genes_ids

rset_go_significant = pd.DataFrame(li, columns = ['GO_term', 'category', 'name', 'level',  'hits_bg', 
                                                  'hits_sample', 'pv_adj_fdr', 'NCBIgeneIDs', 'gene_symbols'])

rset_go_significant.to_csv('revision/Supplementary_tables/RiboSET_ext_GO.txt', sep='\t', index=False)

rset_go_significant[rset_go_significant['hits_sample'] != '0/3304'].groupby(by='category').count()

a = rset_go_significant[rset_go_significant['hits_sample'] != '0/3304']

nucleoplasm = a[a['name'] == 'nucleoplasm'].iloc[0].gene_symbols.split(', ')

cytoplasm = a[a['name'] == 'cytoplasm'].iloc[0].gene_symbols.split(', ')

membrane = a[a['name'] == 'membrane'].iloc[0].gene_symbols.split(', ')

len(set(nucleoplasm).intersection(cytoplasm))

len(set(membrane).intersection(cytoplasm))

len(set(nucleoplasm).intersection(membrane))

len(nucleoplasm), len(membrane), len(cytoplasm)
